+++ changelog.md
@@ -1,6 +1,21 @@
 # Version HEAD
 
+* BREAKING CHANGE: The `Data.Money` module was renamed to `Money`.
+
+* BREAKING CHANGE: The `Data.Money.Internal` module is not exposed anymore. All
+  of its contents are now exported from the `Money` module (with some renaming
+  like `mkDiscreteRep` being now called `fromRawDiscreteRep`, etc.)
+
+* BREAKING CHANGE: Renamed `discreteRep` to `toDiscreteRep`.
+
+* BREAKING CHANGE: Renamed `denseRep` to `toDenseRep`.
+
+* BREAKING CHANGE: Renamed `exchangeRateRep` to `toExchangeRateRep`.
+
 * Remove upper bound constraints from all dependencies except `base`.
+
+* Made dependency on `store` optional for the test suite, so that it can run on
+  GHCJS.
 
 
 # Version 0.2